+++ Sources/SwiftExtras/Extensions/Text+Gradient.swift
@@ -12,6 +12,19 @@
 #if canImport(SwiftUI)
 import SwiftUI
 
+struct EnvironmentWrapperIsEnabled<Content: View>: View {
+    @Environment(\.isEnabled) var isEnabled
+    let content: (Bool) -> Content
+
+    init(@ViewBuilder content: @escaping (Bool) -> Content) {
+        self.content = content
+    }
+
+    var body: some View {
+        content(isEnabled)
+    }
+}
+
 extension Text {
     /// Make the foreground gradient
     ///
@@ -21,28 +34,40 @@ extension Text {
     ///   - colors: Colors
     ///   - startPoint: Startpoint
     ///   - endPoint: End point
-    ///   - if: If condition has met
+    ///   - always: Does it need to ignore an disabled state
     /// - Returns: self
     @ViewBuilder
     public func foregroundLinearGradient(
         colors: [Color] = [.red, .blue, .green, .yellow],
         startPoint: UnitPoint = .leading,
         endPoint: UnitPoint = .trailing,
-        if condition: Bool = true
+        always: Bool = false
     ) -> some View {
-        if condition {
-            self.overlay {
-                LinearGradient(
-                    colors: colors,
-                    startPoint: startPoint,
-                    endPoint: endPoint
-                )
-                .mask(
-                    self
-                )
+        EnvironmentWrapperIsEnabled { isEnabled in
+            if isEnabled || always {
+                if #available(iOS 17, *) {
+                    self.foregroundStyle(
+                        LinearGradient(
+                            colors: colors,
+                            startPoint: startPoint,
+                            endPoint: endPoint
+                        )
+                    )
+                } else {
+                    self.overlay {
+                        LinearGradient(
+                            colors: colors,
+                            startPoint: startPoint,
+                            endPoint: endPoint
+                        )
+                        .mask(
+                            self
+                        )
+                    }
+                }
+            } else {
+                self
             }
-        } else {
-            self
         }
     }
 }